requirements.txt
PyYAML==6.0.1
pandas==2.2.2
numpy==1.26.4
SQLAlchemy==2.0.30
pymysql==1.1.1
streamlit==1.35.0
plotly==5.22.0
seaborn==0.13.2
matplotlib==3.8.4
